ref(search): Extract shared Ask Seer combobox base
Move the common combobox interactions and presentation into a reusable component while keeping the mutation-backed adapter focused on query setup.

Refs BROWSE-628
